Computing (trend-)Limma cell identity genes

exprsMat accepts various matrix objects, including DelayedArray and HDF5Array for out-of-memory computations.

doLimma(exprsMat, cellTypes)

doVoom(exprsMat, cellTypes)

doTtest(exprsMat, cellTypes)

doWilcoxon(exprsMat, cellTypes)

Arguments

exprsMat
expression matrix where columns denote cells and rows denote genes

cellTypes
vector of cell type labels

Value

Returns a list, each element corresponds to statistics (and p-values) associated with a cell-type

Examples

set.seed(1234)
n = 50 ## genes, rows
p = 100 ## cells, cols
exprsMat = matrix(rpois(n*p, lambda = 5), nrow = n)
rownames(exprsMat) = paste0("gene", 1:n)
colnames(exprsMat) = paste0("cell", 1:p)
cellTypes = sample(letters[1:3], size = p, replace = TRUE)

doLimma(exprsMat = exprsMat, cellTypes = cellTypes)

#> $stats
#> DataFrame with 50 rows and 3 columns
#>                a         b         c
#>        <numeric> <numeric> <numeric>
#> gene46   2.97407 -1.433649 -1.849752
#> gene38   2.50385 -3.848203  1.258802
#> gene4    1.61449 -0.878036 -0.894288
#> gene12   1.15447 -2.027285  0.859346
#> gene42   1.13693 -0.302714 -0.968639
#> ...          ...       ...       ...
#> gene48  -1.34354  1.237486  0.206975
#> gene30  -1.73579  0.965303  0.943891
#> gene3   -1.90682  1.778445  0.254905
#> gene13  -2.10457  0.136555  2.248960
#> gene2   -2.67699  1.999426  0.892900
#> 
#> $pvalues
#> DataFrame with 50 rows and 3 columns
#>                 a           b         c
#>         <numeric>   <numeric> <numeric>
#> gene46 0.00295315 0.151736251 0.0644096
#> gene2  0.00745347 0.045617377 0.3719546
#> gene38 0.01231737 0.000120501 0.2081618
#> gene13 0.03537923 0.891388334 0.0245593
#> gene3  0.05660281 0.075392788 0.7988072
#> ...           ...         ...       ...
#> gene47   0.847429    0.276622  0.343298
#> gene20   0.877029    0.897556  0.751763
#> gene29   0.897492    0.014288  0.012634
#> gene15   0.953845    0.523411  0.534271
#> gene11   0.989177    0.605761  0.588453
#> 
#> attr(,"class")
#> [1] "Cepo" "list"

doVoom(exprsMat = exprsMat, cellTypes = cellTypes)

#> $stats
#> DataFrame with 50 rows and 3 columns
#>                a         b         c
#>        <numeric> <numeric> <numeric>
#> gene38   3.22865 -3.762820  0.789522
#> gene46   2.51133 -0.757610 -1.740235
#> gene12   1.65201 -1.944886  0.446500
#> gene42   1.53896 -0.407809 -1.129668
#> gene28   1.40019 -0.533687 -0.972170
#> ...          ...       ...       ...
#> gene13  -1.46087  0.134548  1.736092
#> gene48  -1.70813  1.396626  0.567204
#> gene30  -1.80866  1.300439  0.851046
#> gene3   -1.93852  1.755922  0.377378
#> gene2   -2.87339  2.588534  1.030985
#> 
#> $pvalues
#> DataFrame with 50 rows and 3 columns
#>                 a           b         c
#>         <numeric>   <numeric> <numeric>
#> gene38 0.00136094 0.000184992 0.4302517
#> gene2  0.00430779 0.009878904 0.3031400
#> gene46 0.01296763 0.449937260 0.0835045
#> gene3  0.05336082 0.079626582 0.7060826
#> gene30 0.07135867 0.193964079 0.3952278
#> ...           ...         ...       ...
#> gene22   0.698466    0.982069  0.671942
#> gene11   0.862398    0.934108  0.979052
#> gene21   0.857631    0.634157  0.669791
#> gene31   0.838401    0.685058  0.459046
#> gene45   0.951801    0.220813  0.154378
#> 
#> attr(,"class")
#> [1] "Cepo" "list"

doTtest(exprsMat = exprsMat, cellTypes = cellTypes)

#> $stats
#> DataFrame with 50 rows and 3 columns
#>                a         b         c
#>        <numeric> <numeric> <numeric>
#> gene46   2.69377 -1.382879 -1.727074
#> gene38   2.63884 -4.457442  1.102722
#> gene4    1.38658 -0.936831 -0.828345
#> gene12   1.15694 -2.048086  0.796544
#> gene42   1.06871 -0.283674 -0.911919
#> ...          ...       ...       ...
#> gene48  -1.35447  1.360485  0.225323
#> gene30  -1.66358  0.987731  0.834387
#> gene3   -1.98310  1.714802  0.287203
#> gene13  -2.17495  0.136866  2.241342
#> gene2   -3.00519  2.359443  0.907096
#> 
#> $pvalues
#> DataFrame with 50 rows and 3 columns
#>                 a         b         c
#>         <numeric> <numeric> <numeric>
#> gene1  0.82014854 0.4987028  0.319176
#> gene2  0.00342977 0.0210168  0.370192
#> gene3  0.05030550 0.0923911  0.775140
#> gene4  0.16999352 0.3513082  0.412038
#> gene5  0.76586329 0.6741322  0.428211
#> ...           ...       ...       ...
#> gene46 0.00856449  0.170947 0.0912861
#> gene47 0.83030380  0.225841 0.2803661
#> gene48 0.17945582  0.177977 0.8227355
#> gene49 0.36073112  0.551321 0.7262301
#> gene50 0.61107129  0.302269 0.5906685
#> 
#> attr(,"class")
#> [1] "Cepo" "list"

doWilcoxon(exprsMat = exprsMat, cellTypes = cellTypes)

#> $stats
#> DataFrame with 50 rows and 3 columns
#>                a         b         c
#>        <numeric> <numeric> <numeric>
#> gene2     1605.5     818.0     827.5
#> gene13    1560.5    1051.0     639.5
#> gene30    1480.0     903.0     868.0
#> gene3     1474.5     907.5     869.0
#> gene48    1434.0     889.5     927.5
#> ...          ...       ...       ...
#> gene37    1085.5      1214     951.5
#> gene12    1067.0      1332     852.0
#> gene4     1038.5      1166    1046.5
#> gene46     884.5      1229    1137.5
#> gene38     855.0      1586     810.0
#> 
#> $pvalues
#> DataFrame with 50 rows and 3 columns
#>                 a         b         c
#>         <numeric> <numeric> <numeric>
#> gene1  0.80835248 0.4995051  0.313076
#> gene2  0.00739183 0.0435018  0.376874
#> gene3  0.07983696 0.1781163  0.583678
#> gene4  0.18984748 0.5631293  0.383188
#> gene5  0.77007595 0.6892376  0.441901
#> ...           ...       ...       ...
#> gene46  0.0168803  0.295404  0.109477
#> gene47  0.8681743  0.362425  0.240018
#> gene48  0.1426096  0.138547  0.938973
#> gene49  0.2932627  0.476136  0.667185
#> gene50  0.3982380  0.215740  0.716075
#> 
#> attr(,"class")
#> [1] "Cepo" "list"