Plot densities

Plot densities

plot_densities(
  x,
  dx_list,
  n_genes = NULL,
  assay = "logcounts",
  celltype_col,
  celltype,
  genes = NULL,
  hist_plot = TRUE,
  color = NULL
)

Arguments

x
a SummarizedExperiment or a SingleCellExperiment object.

dx_list
a list containing the ordered output of Dx runs.

assay
a character ("logcounts" by default), indicating the name of the assays(x) element which stores the expression data (i.e., assays(x)$name_assays_expression). We strongly encourage using normalized data, such as counts per million (CPM) or log-CPM.

celltype_col
X.

celltype
a character, indicating the name of the cluster to plot. Default is all clusters.

color
a named color vector.

gene
a character, indicating the name of the gene to plot. Default is NULL. If is.null(gene), this argument overrides the dx_list.

hist
a logical, indicating whether to plot histogram (if TRUE) or not (if FALSE).

Value

A ggplot object.

Author

Hani Jieun Kim hani.kim@sydney.edu.au

plot_density returns a ggplot object with cell-type specific densities for a gene.

Examples

data("cellbench", package = "Cepo")
cellbench

#> class: SingleCellExperiment 
#> dim: 894 895 
#> metadata(3): scPipe Biomart log.exprs.offset
#> assays(2): counts logcounts
#> rownames(894): AP000902.1 TNNI3 ... SCMH1 IGF2BP2
#> rowData names(0):
#> colnames(895): CELL_000001 CELL_000003 ... CELL_000955 CELL_000965
#> colData names(17): unaligned aligned_unmapped ... sizeFactor celltype
#> reducedDimNames(0):
#> altExpNames(0):

ds_res <- Cepo(logcounts(cellbench), cellbench$celltype)

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rowMeans2': could not find function "logcounts"

plot_densities(x = cellbench,
               dx_list = ds_res, 
               n_genes = 2,
               assay = "logcounts", 
               celltype_col = "celltype",
               celltype = names(ds_res))

#> Error in stopifnot((is(x, "SummarizedExperiment") | is(x, "SingleCellExperiment")),     is.numeric(n_genes), length(n_genes) == 1L, is.character(assay),     length(assay) == 1L, is.character(celltype_col), length(celltype_col) ==         1L, is.character(celltype), length(celltype) > 0L, is.logical(hist_plot),     length(hist_plot) == 1L): object 'ds_res' not found

plot_densities(x = cellbench,
               dx_list=ds_res,
               genes = unlist(lapply(ds_res, function(x) names(x)[10:11])),
               assay = "logcounts", 
               celltype_col = "celltype",
               celltype = names(ds_res))

#> Error in stopifnot((is(x, "SummarizedExperiment") | is(x, "SingleCellExperiment")),     is.numeric(n_genes), length(n_genes) == 1L, is.character(assay),     length(assay) == 1L, is.character(celltype_col), length(celltype_col) ==         1L, is.character(celltype), length(celltype) > 0L, is.logical(hist_plot),     length(hist_plot) == 1L): object 'ds_res' not found